% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.h5_features}
\alias{.h5_features}
\title{Feature names stored in a Matilda .h5 (matrix/features).}
\usage{
.h5_features(path)
}
\description{
Feature names stored in a Matilda .h5 (matrix/features).
}
\keyword{internal}
