% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\name{.have_env}
\alias{.have_env}
\title{Is the basilisk matilda environment available (built or buildable)?}
\usage{
.have_env()
}
\description{
Is the basilisk matilda environment available (built or buildable)?
}
\keyword{internal}
