% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pkg_py}
\alias{.pkg_py}
\title{Path to the vendored python package shipped in inst/python/matilda.}
\usage{
.pkg_py()
}
\description{
Path to the vendored python package shipped in inst/python/matilda.
}
\keyword{internal}
