% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{.read_latent}
\alias{.read_latent}
\title{Read the latent embedding (latent_space.csv): cells x z_dim matrix.}
\usage{
.read_latent(path)
}
\description{
Read the latent embedding (latent_space.csv): cells x z_dim matrix.
}
\keyword{internal}
