% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{.read_sim}
\alias{.read_sim}
\title{Read simulated + real cells (sim_data_*.csv / real_data_*.csv / *_label.csv).}
\usage{
.read_sim(dir)
}
\value{
list(sim = list(rna,adt,atac,label), real = list(rna,adt,atac,label))
}
\description{
Read simulated + real cells (sim_data_*.csv / real_data_*.csv / *_label.csv).
}
\keyword{internal}
