% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{.slice_to_features}
\alias{.slice_to_features}
\title{Slice an SCE to the model's exact features + order (drop extra features / altExps).}
\usage{
.slice_to_features(sce, feats, adt_exp = "ADT", atac_exp = "ATAC")
}
\description{
Slice an SCE to the model's exact features + order (drop extra features / altExps).
}
\keyword{internal}
