% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{.train_mods}
\alias{.train_mods}
\title{Write modalities to temp files and call the path-based trainer.}
\usage{
.train_mods(mods, hp)
}
\description{
Write modalities to temp files and call the path-based trainer.
}
\keyword{internal}
