% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matilda-package.R
\docType{package}
\name{matilda-package}
\alias{matilda}
\alias{matilda-package}
\title{matilda: R interface to Matilda multi-task single-cell multimodal learning}
\description{
Trains a multimodal VAE + classifier (RNA / ADT / ATAC) and runs
classification, dimension reduction, feature selection and simulation. The
upstream PyTorch implementation is vendored under \code{inst/python/matilda}
and driven via \pkg{reticulate} inside a \pkg{basilisk} environment. Results
are written back into the input \link[SingleCellExperiment]{SingleCellExperiment}.
}
\author{
\strong{Maintainer}: Sichang Ding \email{usyddsc@gmail.com}

Authors:
\itemize{
  \item Chunlei Liu \email{chunlei@cmri.com}
  \item Pengyi Yang
}

}
\keyword{internal}
