% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matilda_example_sce}
\alias{matilda_example_sce}
\title{A tiny multimodal SingleCellExperiment for examples and quick trials.}
\usage{
matilda_example_sce(n_cells = 60)
}
\arguments{
\item{n_cells}{number of cells.}
}
\value{
a SingleCellExperiment with altExps "ADT"/"ATAC" and a colData "cell_type".
}
\description{
Builds a small synthetic RNA + ADT + ATAC \link[SingleCellExperiment]{SingleCellExperiment}
with imbalanced cell types, suitable for trying the Matilda workflow.
}
\examples{
sce <- matilda_example_sce()
sce
}
